@@ -9,8 +9,8 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
         <maven.surefire.plugin.version>3.2.5</maven.surefire.plugin.version>
         <junit.jupiter.version>5.10.2</junit.jupiter.version>
     </properties>
